requirements.txt
pysui>=0.98.0
python-dotenv
